+++ app/utils/screenshots/capture.py
@@ -1,1064 +1,5 @@
-# utils/screenshots.py
-
-import datetime
-import io
-import ipaddress
-import json
-import logging
-import os
-import platform
-import re
-import shutil
-import socket
-import subprocess
-import time
-from urllib.parse import urlparse
-import glob
-import base64
-import nodriver
-import psutil
-from werkzeug.utils import secure_filename
-import urllib3
-from dateutil import tz
-import random
-
-os.environ["WDM_LOG"] = "0"
-os.environ["WDM_LOG_LEVEL"] = "0"
-logging.getLogger("webdriver_manager").setLevel(logging.WARNING)
-logging.getLogger("urllib3").setLevel(logging.ERROR)
-urllib3.disable_warnings(urllib3.exceptions.InsecureRequestWarning)
-
-import numpy as np
-import requests
-import yt_dlp as youtube_dl
-from pdf2image import convert_from_path
-from PIL import (
-    Image,
-    ImageDraw,
-    ImageFont,
-    ImageOps,
-    ImageStat,
-)
-import textwrap
-from pyvirtualdisplay import Display
-from selenium import webdriver
-from selenium.webdriver.chrome.options import Options
-from selenium.webdriver.chrome.service import Service
-from selenium.webdriver.common.by import By
-from selenium.common.exceptions import TimeoutException, WebDriverException
-from webdriver_manager.chrome import ChromeDriverManager
-
-try:
-    from pynput import mouse, keyboard
-except Exception as e:  # pragma: no cover - optional dependency
-    mouse = None
-    keyboard = None
-    logging.warning("pynput not available: %s", e)
-
-
-from app.config import (
-    DEBUG,
-    LANG,
-    SCREENSHOT_DIRECTORY,
-    UA,
-    FFMPEG_PATH,
-    FFMPEG_HWACCEL,
-    NUM_FRAMES,
-    CAPTURE_TIMEOUT,
-    PROBE_SIZE_DEFAULT,
-    PROBE_SIZE_RTSP,
-    PROBE_SIZE_OTHER,
-    ANALYZE_DURATION_DEFAULT,
-    ANALYZE_DURATION_RTSP,
-    ANALYZE_DURATION_OTHER,
-    TZ,
-)
-import app.config as config
-from app.utils.validators import validate_proxy, validate_url
-
-last_camera_test = {}
-last_camera_test_time = {}
-last_camera_header = {}
-last_camera_header_time = {}
-last_camera_light = {}
-last_camera_light_time = {}
-lurl_cache = {}
-lurl_cache_time = {}
-throttle_cache = {}
-chrome_version = {}
-last_modified_cache = {}
-etag_cache = {}
-
-# Cache of last HTTP status codes per URL
-status_code_cache = {}
-status_code_cache_time = {}
-STATUS_CACHE_TTL = 60 * 60  # 1 hour
-STATUS_CACHE_PATH = "data/status_cache.json"
-
-
-def _load_status_cache() -> None:
-    """Load cached status codes from ``STATUS_CACHE_PATH``."""
-    if not os.path.exists(STATUS_CACHE_PATH):
-        return
-    try:
-        with open(STATUS_CACHE_PATH, "r") as f:
-            data = json.load(f)
-    except Exception:
-        return
-
-    status_code_cache.clear()
-    status_code_cache_time.clear()
-    for url, info in data.items():
-        status_code_cache[url] = info.get("code")
-        status_code_cache_time[url] = info.get("time", 0)
-
-
-def _persist_status_cache() -> None:
-    """Write ``status_code_cache`` to ``STATUS_CACHE_PATH``."""
-    os.makedirs(os.path.dirname(STATUS_CACHE_PATH), exist_ok=True)
-    data = {
-        url: {"code": code, "time": status_code_cache_time.get(url, 0)}
-        for url, code in status_code_cache.items()
-    }
-    try:
-        with open(STATUS_CACHE_PATH, "w") as f:
-            json.dump(data, f)
-    except Exception:
-        logging.exception("Failed to persist status cache")
-
-
-_load_status_cache()
-
-
-FONT_CANDIDATES = [
-    "DejaVuSans-Bold.ttf",
-    "DejaVuSans.ttf",
-    "Arial.ttf",
-    "LiberationSans-Regular.ttf",
-]
-
-# User agent templates used when stealth mode is enabled. The actual version is
-# filled dynamically based on the installed Chrome version so that outdated
-# strings are avoided.
-STEALTH_UA_TEMPLATES = [
-    (
-        "Mozilla/5.0 (Windows NT 10.0; Win64; x64) "
-        "AppleWebKit/537.36 (KHTML, like Gecko) Chrome/{version}.0.0.0 "
-        "Safari/537.36"
-    ),
-    (
-        "Mozilla/5.0 (Macintosh; Intel Mac OS X 13_4) "
-        "AppleWebKit/537.36 (KHTML, like Gecko) Chrome/{version}.0.0.0 "
-        "Safari/537.36"
-    ),
-    (
-        "Mozilla/5.0 (X11; Linux x86_64) "
-        "AppleWebKit/537.36 (KHTML, like Gecko) Chrome/{version}.0.0.0 "
-        "Safari/537.36"
-    ),
-]
-
-
-def random_user_agent():
-    """Return a randomized user agent string for stealth mode."""
-    chrome_path = get_chrome_path()
-    version = get_chrome_version(chrome_path)
-    # Pick a nearby version to avoid obvious automation patterns
-    major_version = random.randint(max(100, version - 1), version + 1)
-    template = random.choice(STEALTH_UA_TEMPLATES)
-    return template.format(version=major_version)
-
-
-def load_font(size):
-    """Return a truetype font for overlays."""
-    for font_name in FONT_CANDIDATES:
-        try:
-            return ImageFont.truetype(font_name, size)
-        except IOError:
-            continue
-    return ImageFont.load_default()
-
-
-# Global flag to track user activity
-user_active = False
-
-_DRIVER = None
-
-
-def get_driver(opts):
-    global _DRIVER
-    if _DRIVER is None:
-        service = Service(ChromeDriverManager().install())
-        _DRIVER = webdriver.Chrome(service=service, options=opts)
-    return _DRIVER
-
-
-_session = None
-
-
-def http_session():
-    global _session
-    if _session is None:
-        _session = requests.Session()
-        _session.verify = False
-        _session.headers.update({"user-agent": UA})
-        _session.headers.update({"Accept": "*/*"})
-        _session.mount("http://", requests.adapters.HTTPAdapter(pool_maxsize=20))
-        _session.mount("https://", requests.adapters.HTTPAdapter(pool_maxsize=20))
-    return _session
-
-
-def _is_valid_png(path):
-    try:
-        with Image.open(path) as im:
-            im.verify()  # raises if corrupt/zero-byte
-        return True
-    except Exception:
-        return False
-
-
-def get_cached_status_code(url):
-    """Return cached HTTP status code for URL if not expired."""
-    code = status_code_cache.get(url)
-    ts = status_code_cache_time.get(url, 0)
-    if code is not None and time.time() - ts < STATUS_CACHE_TTL:
-        return code
-    if code is not None:
-        # Entry expired, remove and persist cleanup
-        status_code_cache.pop(url, None)
-        status_code_cache_time.pop(url, None)
-        _persist_status_cache()
-    return None
-
-
-def set_cached_status_code(url, code):
-    """Store status code for URL with current timestamp."""
-    status_code_cache[url] = code
-    status_code_cache_time[url] = time.time()
-    _persist_status_cache()
-
-
-# Callback functions to update activity state
-def on_move(x, y):
-    global user_active
-    user_active = True
-
-
-def on_click(x, y, button, pressed):
-    global user_active
-    user_active = True
-
-
-def on_scroll(x, y, dx, dy):
-    global user_active
-    user_active = True
-
-
-def on_press(key):
-    global user_active
-    user_active = True
-
-
-import ctypes
-import ctypes.util
-import os
-import threading
-import time
-
-_idle_lock = threading.Lock()
-_x11 = None
-_xss = None
-
-
-class XScreenSaverInfo(ctypes.Structure):
-    _fields_ = [
-        ("window", ctypes.c_ulong),
-        ("state", ctypes.c_int),
-        ("kind", ctypes.c_int),
-        ("since", ctypes.c_ulong),  # ms since state started
-        ("idle", ctypes.c_ulong),  # ms idle (what we need)
-        ("eventMask", ctypes.c_ulong),
-    ]
-
-
-def idle_seconds_x11() -> int:
-    """Return idle seconds on X11 systems."""
-
-    dpy_name = os.environ.get("DISPLAY")
-    if not dpy_name:
-        raise RuntimeError("$DISPLAY is not set – not running under X11.")
-
-    global _x11, _xss
-
-    with _idle_lock:
-        if _x11 is None or _xss is None:
-            libX11_path = ctypes.util.find_library("X11")
-            libXss_path = ctypes.util.find_library("Xss")
-            if not (libX11_path and libXss_path):
-                raise RuntimeError(
-                    "libX11 or libXss not found (install libx11-6 libxss1)."
-                )
-
-            _x11 = ctypes.cdll.LoadLibrary(libX11_path)
-            _xss = ctypes.cdll.LoadLibrary(libXss_path)
-
-            _x11.XOpenDisplay.argtypes = [ctypes.c_char_p]
-            _x11.XOpenDisplay.restype = ctypes.c_void_p
-            _x11.XDefaultRootWindow.argtypes = [ctypes.c_void_p]
-            _x11.XDefaultRootWindow.restype = ctypes.c_ulong
-            _xss.XScreenSaverAllocInfo.restype = ctypes.POINTER(XScreenSaverInfo)
-            _xss.XScreenSaverQueryInfo.argtypes = [
-                ctypes.c_void_p,
-                ctypes.c_ulong,
-                ctypes.POINTER(XScreenSaverInfo),
-            ]
-            _xss.XScreenSaverQueryInfo.restype = ctypes.c_int
-            _x11.XFree.argtypes = [ctypes.c_void_p]
-            _x11.XFree.restype = None
-            _x11.XCloseDisplay.argtypes = [ctypes.c_void_p]
-            _x11.XCloseDisplay.restype = None
-
-        x11 = _x11
-        xss = _xss
-
-        dpy = x11.XOpenDisplay(dpy_name.encode())
-        if not dpy:
-            raise RuntimeError(f"cannot open X display '{dpy_name}'")
-
-        info = xss.XScreenSaverAllocInfo()
-        if not info:
-            x11.XCloseDisplay(dpy)
-            raise RuntimeError("XScreenSaverAllocInfo returned NULL")
-
-        root = x11.XDefaultRootWindow(dpy)
-        status = xss.XScreenSaverQueryInfo(dpy, root, info)
-        if status == 0:
-            x11.XFree(info)
-            x11.XCloseDisplay(dpy)
-            raise RuntimeError("XScreenSaver extension not active on this X server")
-
-        idle_ms = info.contents.idle
-        x11.XFree(info)
-        x11.XCloseDisplay(dpy)
-        return idle_ms // 1000
-
-
-def idle_seconds_loginctl() -> int:
-    """Return seconds of user idleness according to systemd-logind.
-    0  → actively using keyboard/mouse right now."""
-    import os, subprocess, time
-
-    uid = os.getuid()
-    try:
-        out = subprocess.check_output(
-            [
-                "loginctl",
-                "show-user",
-                str(uid),
-                "-p",
-                "IdleHint",
-                "-p",
-                "IdleSinceHintMonotonicUSec",
-            ],
-            text=True,
-            timeout=0.3,  # fail fast
-        ).splitlines()
-    except subprocess.SubprocessError:
-        raise RuntimeError("loginctl unavailable")
-
-    props = dict(l.split("=", 1) for l in out if "=" in l)
-    if props.get("IdleHint", "no") != "yes":
-        return 0  # user is active
-
-    idle_us = int(props["IdleSinceHintMonotonicUSec"])
-    return int((time.monotonic() * 1_000_000 - idle_us) / 1_000_000)
-
-
-# Function to detect user activity
-def check_user_activity(timeout=10):
-
-    global user_active
-    user_active = False
-
-    # oiq = make_idle_irq()
-    # liq = oiq()
-    # if liq < 120:
-    #    user_active = True  # allow to check on listeners for the 0 second case
-    #    return user_active
-    try:
-        idle_seconds_x = idle_seconds_x11()
-        if idle_seconds_x < 120:
-            user_active = True  # user recently active
-            return user_active
-    except Exception as e:
-        logging.debug(f"idle_seconds_x11 failed: {e}")
-
-    # idle_seconds = idle_seconds_loginctl()
-    # if 1 < idle_seconds < 120:
-    #    user_active = True  # allow to check on listeners for the 0 second case
-    #    return user_active
-
-    if mouse is None or keyboard is None:
-        return user_active
-
-    # Create listeners for keyboard and mouse
-    mouse_listener = mouse.Listener(
-        on_move=on_move, on_click=on_click, on_scroll=on_scroll
-    )
-    keyboard_listener = keyboard.Listener(on_press=on_press)
-
-    # Start listeners
-    mouse_listener.start()
-    keyboard_listener.start()
-
-    # Monitor for a defined timeout
-    start_time = time.time()
-    while time.time() - start_time < timeout:
-        if user_active:
-            break
-        time.sleep(0.1)
-
-    # Stop listeners
-    mouse_listener.stop()
-    keyboard_listener.stop()
-
-    # Ensure threads close their X connections before returning
-    mouse_listener.join()
-    keyboard_listener.join()
-
-    return user_active
-
-
-def _send_input_event():
-    """Move the mouse slightly to generate an input event."""
-    if mouse is None:
-        return
-    try:
-        controller = mouse.Controller()
-        x, y = controller.position
-        controller.move(1, 0)
-        controller.move(-1, 0)
-        controller.position = (x, y)
-    except Exception as e:  # pragma: no cover - best effort
-        logging.debug(f"_send_input_event failed: {e}")
-
-
-def detect_background_color(image: Image.Image, sample_width: int = 10):
-    """Return the most common color found along the image border."""
-    arr = np.asarray(image.convert("RGBA"))
-    top = arr[:sample_width, :, :].reshape(-1, 4)
-    bottom = arr[-sample_width:, :, :].reshape(-1, 4)
-    left = arr[:, :sample_width, :].reshape(-1, 4)
-    right = arr[:, -sample_width:, :].reshape(-1, 4)
-    border = np.concatenate([top, bottom, left, right], axis=0)
-    colors, counts = np.unique(border, axis=0, return_counts=True)
-    return tuple(int(c) for c in colors[counts.argmax()])
-
-
-def remove_background(image, background_color=None, threshold=10):
-    """Crop the image to remove the background color border and ensure a 16:9 aspect ratio."""
-    if background_color is None:
-        background_color = detect_background_color(image)
-    # Find the bounding box of the non-background area
-    bbox = find_bounding_box(image, background_color, threshold)
-
-    # Adjust the bounding box to fit a 16:9 aspect ratio
-    if bbox:
-        bbox = adjust_bbox_to_aspect_ratio(bbox, image.size, aspect_ratio=(16, 9))
-        image = image.crop(bbox)
-    return image
-
-
-def find_bounding_box(
-    image: Image.Image,
-    background_color=(14, 14, 14, 255),
-    threshold: int = 10,
-):
-    """
-    Return (left, top, right, bottom) that contains all pixels whose
-    per-channel distance from `background_color` > threshold.
-
-    If *every* pixel is background, returns None.
-    """
-    # RGBA → ndarray(H, W, 4)
-    arr = np.asarray(image.convert("RGBA"), dtype=np.int16)
-
-    bg = np.array(background_color, dtype=np.int16)
-    # True where *any* channel differs more than threshold
-    fg_mask = np.any(np.abs(arr - bg) > threshold, axis=-1)
-
-    if not fg_mask.any():  # all background
-        return None
-
-    ys, xs = np.nonzero(fg_mask)
-    top, bottom = ys.min(), ys.max()
-    left, right = xs.min(), xs.max()
-
-    return int(left), int(top), int(right), int(bottom)
-
-
-def adjust_bbox_to_aspect_ratio(bbox, image_size, aspect_ratio=(16, 9)):
-    """Adjust the bounding box to fit the specified aspect ratio."""
-    left, top, right, bottom = bbox
-    bbox_width = right - left
-    bbox_height = bottom - top
-    if bbox_height == 0:
-        bbox_height = 1
-
-    bbox_aspect_ratio = bbox_width / bbox_height
-
-    target_aspect_ratio = aspect_ratio[0] / aspect_ratio[1]
-
-    if bbox_aspect_ratio > target_aspect_ratio:
-        # The bounding box is too wide, adjust the height
-        new_height = bbox_width / target_aspect_ratio
-        vertical_padding = (new_height - bbox_height) / 2
-        top = max(0, top - vertical_padding)
-        bottom = min(image_size[1], bottom + vertical_padding)
-    elif bbox_aspect_ratio < target_aspect_ratio:
-        # The bounding box is too tall, adjust the width
-        new_width = bbox_height * target_aspect_ratio
-        horizontal_padding = (new_width - bbox_width) / 2
-        left = max(0, left - horizontal_padding)
-        right = min(image_size[0], right + horizontal_padding)
-
-    return (int(left), int(top), int(right), int(bottom))
-
-
-def is_similar_color(color1, color2, threshold):
-    """Check if two colors are similar."""
-    return all(abs(c1 - c2) <= threshold for c1, c2 in zip(color1, color2))
-
-
-def is_mostly_blank(
-    image: Image.Image,
-    threshold: float = 0.98,
-    blank_color=(255, 255, 255),
-    text_std_threshold: int = 20,
-    dark_threshold: int = 10,
-):
-    """
-    True  → we consider the frame “uninteresting” (blank / flat / too dark).
-    False → keep the frame.
-    """
-    arr = np.asarray(image.convert("RGB"), dtype=np.int16)
-
-    # ---------- 1.  “Mostly blank?”  ----------
-    blank = np.array(blank_color, dtype=np.int16)
-    blank_px = np.all(np.abs(arr - blank) <= 30, axis=-1).mean()
-    if blank_px >= threshold:
-        return True
-
-    # ---------- 2.  “Flat image?”  ----------
-    # Low global std-dev ≈ little structure / shapes
-    if arr.std() < text_std_threshold:
-        return True
-
-    # ---------- 3.  “Too dark?”  ----------
-    # Use perceptual luma so pure-dark blue isn’t mis-treated
-    luma = np.dot(arr.mean(axis=(0, 1)), [0.2126, 0.7152, 0.0722])
-    if luma < dark_threshold:
-        return True
-
-    return False
-
-
-def run_cmd(cmd, timeout):
-    proc = subprocess.Popen(cmd, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
-    try:
-        out, err = proc.communicate(timeout=timeout)
-    except subprocess.TimeoutExpired:
-        proc.kill()  # SIGKILL
-        out, err = proc.communicate()
-        raise RuntimeError(f"timeout: {' '.join(cmd)}")
-    if proc.returncode:
-        raise RuntimeError(err.decode()[:300])
-    return out
-
-
-def add_timestamp(image_path, name="unknown", invert=False):
-    if os.path.exists(image_path):
-        with Image.open(
-            image_path
-        ) as image:  # consider unlinking if this fails to open
-            # Convert the image to RGBA mode in case it's a format that doesn't support transparency
-            try:
-                image = image.convert("RGB")
-            except Exception as e:
-                # for debugging only, otherwise unlink the file
-                if DEBUG:
-                    os.rename(image_path, image_path.replace(".png", ".broken"))
-                else:
-                    # unlink the offending image
-                    os.unlink(image_path)
-                # print(" warning : image load issue:", image_path, e)
-                logging.error(f"Error saving image: {image_path} {e}")
-                return
-
-            draw = ImageDraw.Draw(image)
-            # If the image has the "invert" flag, then invert colors for readability
-            if invert:
-                image = ImageOps.invert(image)
-
-            # Define the timestamp format
-
-            zone = tz.gettz(TZ) or tz.UTC  # fall back if the name is invalid
-            local_time = datetime.datetime.now(zone)
-            tz_name = local_time.tzname() or ""
-            timestamp = local_time.strftime("%Y-%m-%d %H:%M:%S")
-            if tz_name:
-                timestamp = f"{timestamp} {tz_name}"
-
-            utc_time = datetime.datetime.utcnow()
-            utc_timestamp = utc_time.strftime("%Y-%m-%d %H:%M:%S UTC")
-
-            max_height = min(image.height, image.width * 9 // 16)
-            font_size = int(max_height * 0.05)
-            if font_size < 5:  # ignore tiny images
-                return
-
-            top_offset = (image.height - max_height) / 2
-
-            # Use the helper to load fonts. The small font is half-sized.
-            font = load_font(font_size)
-            font_small = load_font(int(max(5, font_size / 2)))
-
-            padding = 6
-
-            # Render the name in the upper-left corner
-            x = padding
-            y = int(padding + top_offset)
-            bbox = draw.textbbox((x, y), name, font=font, stroke_width=1)
-            text_w = bbox[2] - bbox[0]
-            text_h = bbox[3] - bbox[1]
-            background = Image.new(
-                "RGBA",
-                (text_w + padding * 2, text_h + padding * 2),
-                (0, 0, 0, 128),
-            )
-            image.paste(background, (x - padding, y - padding), background)
-            draw.text(
-                (x, y),
-                name,
-                font=font,
-                fill=(255, 255, 255, 255),
-                stroke_width=1,
-                stroke_fill=(0, 0, 0, 255),
-            )
-
-            # Timestamp in the lower-right corner
-            text_w = draw.textbbox((0, 0), timestamp, font=font, stroke_width=1)[2]
-            x = image.width - text_w - padding
-            y = int(image.height - top_offset - font_size * 2)
-            bbox = draw.textbbox((x, y), timestamp, font=font, stroke_width=1)
-            text_w = bbox[2] - bbox[0]
-            text_h = bbox[3] - bbox[1]
-            background = Image.new(
-                "RGBA",
-                (text_w + padding * 2, text_h + padding * 2),
-                (0, 0, 0, 128),
-            )
-            image.paste(background, (bbox[0] - padding, bbox[1] - padding), background)
-
-            draw.text(
-                (x, y),
-                timestamp,
-                font=font,
-                fill=(255, 255, 255, 255),
-                stroke_width=1,
-                stroke_fill=(0, 0, 0, 255),
-            )
-
-            if utc_timestamp != timestamp:
-                tz_bbox = draw.textbbox(
-                    (0, 0), utc_timestamp + "Z", font=font_small, stroke_width=1
-                )
-                text_w = tz_bbox[2] - tz_bbox[0]
-                text_h = tz_bbox[3] - tz_bbox[1]
-                background = Image.new(
-                    "RGBA",
-                    (text_w + padding * 2, text_h + padding * 2),
-                    (0, 0, 0, 128),
-                )
-                tz_x = x
-                tz_y = y + font_size + padding
-                image.paste(
-                    background,
-                    (tz_x - padding, tz_y - padding),
-                    background,
-                )
-                draw.text(
-                    (tz_x, tz_y),
-                    utc_timestamp + "Z",
-                    font=font_small,
-                    fill=(255, 255, 255, 255),
-                    stroke_width=1,
-                    stroke_fill=(0, 0, 0, 255),
-                )
-
-            # Save the image
-            image.save(image_path, "PNG")
-
-        try:
-            from .qrcode_overlay import add_micro_barcode
-
-            add_micro_barcode(image_path, name)
-        except Exception as e:  # pragma: no cover - overlay failures are non-critical
-            logging.debug(f"Micro barcode overlay failed: {e}")
-
-
-def create_placeholder(image_path, name="unknown"):
-    """Generate a simple placeholder image with a timestamp."""
-    img = Image.new("RGB", (640, 360), color="black")
-    draw = ImageDraw.Draw(img)
-    font = load_font(20)
-    zone = tz.gettz(TZ) or tz.UTC
-    timestamp = datetime.datetime.now(zone).strftime("%Y-%m-%d %H:%M:%S")
-    text = f"{name}\n{timestamp}"
-    draw.multiline_text((10, 10), text, fill=(255, 255, 255), font=font)
-    img.save(image_path, "PNG")
-
-
-def download_image(
-    url,
-    output_path,
-    timeout=CAPTURE_TIMEOUT,
-    name="unknown",
-    invert=False,
-    dark=False,
-    stealth=False,
-    proxy=None,
-    username=None,
-    password=None,
-):
-    """Attempt to download an image directly from the URL and convert it to PNG format.
-
-    Args:
-        url (str): Image URL.
-        output_path (str): Where to save the PNG.
-        timeout (int): Timeout in seconds.
-        name (str): Friendly name for logging.
-        invert (bool): If True, invert timestamp colors.
-        dark (bool): Apply dark mode.
-        stealth (bool): Use stealth user agent.
-        proxy (str, optional): Proxy to use for the HTTP request.
-    """
-
-    proxy = validate_proxy(proxy)
-
-    # ideally the timeout should be pretty high, its an image, and it could be real big
-    if timeout < 10:
-        timeout = 10
-
-    response = None
-
-    cached = get_cached_status_code(url)
-    if cached is not None and cached != 200:
-        logging.debug(f"Skipping {url} due to cached status {cached}")
-        return False
-    try:
-        lua = UA
-        if stealth:
-            lua = random_user_agent()
-        headers = {"user-agent": lua}
-        proxies = {"http": proxy, "https": proxy} if proxy else None
-
-        auth = get_auth(url, username, password)
-
-        request_kwargs = dict(
-            stream=True,
-            timeout=(timeout, timeout * 3),
-            verify=False,
-            headers=headers,
-            auth=auth,
-        )
-        if proxies:
-            request_kwargs["proxies"] = proxies
-
-        response = http_session().get(url, **request_kwargs)
-        if response.status_code == 401 and auth is not None:
-            auth = get_digest_auth(url, username, password)
-            request_kwargs = dict(
-                stream=True,
-                timeout=(timeout, timeout * 3),
-                verify=False,
-                headers=headers,
-                auth=auth,
-            )
-            if proxies:
-                request_kwargs["proxies"] = proxies
-
-            response = http_session().get(url, **request_kwargs)
-
-        status = response.status_code
-        set_cached_status_code(url, status)
-
-        if status == 200:
-            # Open the image directly from the response bytes
-            image = Image.open(io.BytesIO(response.content))
-            response.close()
-
-            # Convert the image to RGBA mode in case it's a format that doesn't support transparency
-            image = image.convert("RGB")
-            image = remove_background(image)
-            if dark:
-                apply_dark_mode(image)
-
-            os.makedirs(os.path.dirname(output_path), exist_ok=True)
-            tmp_path = output_path + ".tmp"
-
-            # Save to a temporary file first so readers don't see partial data
-            image.save(tmp_path, "PNG")
-            if os.path.exists(tmp_path) and _is_valid_png(tmp_path):
-                add_timestamp(tmp_path, name=name, invert=invert)
-                os.replace(tmp_path, output_path)
-                return True
-            if os.path.exists(tmp_path):
-                os.remove(tmp_path)
-        else:
-            response.close()
-            logging.warning(f"Error downloading image: HTTP status code {status} {url}")
-    except Exception as e:
-        logging.error(f"Error downloading image: {e} {url} {timeout}")
-        set_cached_status_code(url, 0)
-    finally:
-        if response is not None:
-            response.close()  # Ensure the connection is closed
-
-    return False
-
-
-def download_pdf(
-    url,
-    output_path,
-    timeout=CAPTURE_TIMEOUT,
-    name="unknown",
-    invert=False,
-    dark=False,
-    stealth=False,
-    username=None,
-    password=None,
-):
-    """
-    Attempt to download the first page of a PDF from the URL and convert it to PNG format.
-    """
-    tmp_name = None  # path of the downloaded PDF
-    lsuccess = False
-
-    cached = get_cached_status_code(url)
-    if cached is not None and cached != 200:
-        logging.debug(f"Skipping PDF download for {url} due to cached status {cached}")
-        return False
-
-    if timeout < 10:
-        timeout = 10
-
-    try:
-        # Download the PDF file
-        lua = UA
-        if stealth:
-            lua = random_user_agent()
-
-        headers = {"user-agent": lua}
-        auth = get_auth(url, username, password)
-
-        response = http_session().get(
-            url,
-            stream=True,
-            timeout=timeout,
-            verify=False,
-            headers=headers,
-            auth=auth,
-            allow_redirects=True,
-        )
-
-        # Possibly re-try with DigestAuth if 401
-        if response.status_code == 401 and auth is not None:
-            auth = get_digest_auth(url, username, password)
-            response = http_session().get(
-                url,
-                stream=True,
-                timeout=timeout,
-                verify=False,
-                headers=headers,
-                auth=auth,
-                allow_redirects=True,
-            )
-        set_cached_status_code(url, response.status_code)
-
-        if response.status_code != 200:
-            logging.error(f"Error downloading PDF: HTTP {response.status_code}")
-            return False
-
-        # Save PDF to a temp file
-
-        tmpdirname = f"/tmp/glimpser_{name}"
-        os.makedirs(tmpdirname, exist_ok=True)
-        if os.path.exists(tmpdirname):
-            with open(os.path.join(tmpdirname, "file.pdf"), "wb") as tmp:
-                tmp_name = tmp.name
-                tmp.write(response.content)
-
-        # Convert the first page to an image
-        pages = convert_from_path(tmp_name, first_page=1, last_page=1)
-        if not pages:
-            logging.error("Error converting PDF to image: No pages found")
-            return False
-
-        image = pages[0].convert("RGB")
-        image = remove_background(image)
-        if dark:
-            image = apply_dark_mode(image)
-
-        os.makedirs(os.path.dirname(output_path), exist_ok=True)
-        tmp_path = output_path + ".tmp"
-        image.save(tmp_path, "PNG")
-
-        if os.path.exists(tmp_path) and _is_valid_png(tmp_path):
-            add_timestamp(tmp_path, name=name, invert=invert)
-            os.replace(tmp_path, output_path)
-            logging.debug(f"Successfully saved PDF page to {output_path}")
-            lsuccess = True
-        else:
-            if os.path.exists(tmp_path):
-                os.remove(tmp_path)
-        return lsuccess
-
-    except Exception as e:
-        logging.error(f"Error downloading PDF: {e}")
-        set_cached_status_code(url, 0)
-        return False
-
-    finally:
-        # Always remove leftover PDF, unless you specifically want to keep it
-        if tmp_name and os.path.exists(tmp_name):
-            try:
-                os.remove(tmp_name)
-            except OSError:
-                pass
-
-
-def is_enhanced(url):
-    extractors = youtube_dl.extractor.gen_extractors()
-    for e in extractors:
-        if e.suitable(url) and e.IE_NAME != "generic":
-            return True
-    return False
-
-
-def get_arp_output(ip_address, timeout):
-    if platform.system() == "Windows":
-        command = ["arp", "-a", ip_address]
-    else:
-        command = ["ip", "neigh", "show", ip_address]
-    try:
-        return subprocess.check_output(
-            command, stderr=subprocess.STDOUT, timeout=timeout
-        )
-    except FileNotFoundError:
-        return b""
-
-
-def is_private_ip(ip_address):
-    return ipaddress.ip_address(ip_address).is_private
-
-
-def is_address_reachable(address, port=80, timeout=5):
-
-    if port is None:
-        port = 80
-
-    if timeout < 3:
-        timeout = 3
-
-    try:
-        # Resolve the domain name to an IP address
-        ip_address = socket.gethostbyname(address)
-        # print(f"{address} resolved to {ip_address}")
-    except Exception:
-        if address in ("google.com", "www.google.com"):
-            return True
-        return False
-
-    # check the arp table, particularly if its an unroutable ip address
-    if is_private_ip(ip_address):
-        try:
-            arp_entry = get_arp_output(ip_address, timeout).lower()
-            if "no entry" in arp_entry.decode().lower():
-                logging.warning("failing arp entry for %s", ip_address)
-                return False
-        except Exception as e:
-            logging.warning("failure to arp %s", e)
-
-    try:
-        # Create a socket object
-        sock = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-        sock.settimeout(timeout)
-        # Attempt to connect to the address on the specified port
-        result = sock.connect_ex((ip_address, port))
-        sock.close()
-        if result == 0:
-            # print(f"Successfully connected to {ip_address} on port {port}")
-            # print(f"Failed to connect to {ip_address} on port {port}")
-
-            return True
-        if address in ("google.com", "www.google.com"):
-            return True
-        return False
-    except Exception as e:
-        logging.warning(f"Socket error: {e}")
-        if address in ("google.com", "www.google.com"):
-            return True
-
-    return False
-
-
-def parse_url(url):
-    """Return the domain and port extracted from *url*.
-
-    ``urllib.parse.urlparse`` treats strings without a scheme oddly.  For
-    example ``"example.com:8080/path"`` is parsed with ``"example.com"`` as the
-    *scheme* rather than the hostname.  To handle such URLs we prefix ``"//"`` so
-    they are interpreted as network locations.
-    """
-
-    # Handle URLs missing a scheme like ``example.com:8080/path`` by prefixing
-    # ``//`` which causes ``urlparse`` to parse the hostname and port correctly.
-    if "://" not in url:
-        parsed_url = urlparse("//" + url)
-    else:
-        parsed_url = urlparse(url)
-
-    domain = parsed_url.hostname
-    if parsed_url.scheme == "" and domain is None:
-        domain = re.sub(r"\/.+?$", "", parsed_url.path)
-
-    port = parsed_url.port
-    # If the port is None and the scheme is specified, infer the default port.
-    if port is None:
-        if parsed_url.scheme == "http":
-            port = 80
-        elif parsed_url.scheme == "https":
-            port = 443
-        # Add more schemes and their default ports if necessary.
-
-    return domain, port
-
-
-def cas_error(url):
-
-    entry = throttle_cache.setdefault(url, {"errors": 0, "first": time.time()})
-
-    if entry.get("last", 0) > time.time() - 60 * 5:
-        entry["errors"] += 1
-    else:
-        entry["errors"] = 1
-        entry["first"] = time.time()
-    entry["last"] = time.time()
-
-    if entry["errors"] > 2:
-        logging.error(f"Could not reach host: {url} {entry['errors']} times")
-        entry["timeout"] = time.time() + 60 * 60  # 1 hour timeout
+from .helpers import *
+from .processing import is_mostly_blank, remove_background, create_placeholder
 
 
 def capture_or_download(name: str, template: dict) -> bool:
